Move side panels to new users folder
