+++ kolibri/plugins/facility/assets/src/utils.js
@@ -3,11 +3,11 @@ import { isNavigationFailure, NavigationFailureType } from 'vue-router';
 import logger from 'kolibri-logging';
 import useFacilities from 'kolibri-common/composables/useFacilities';
 import { PageNames } from './constants';
-import MoveToTrashSidePanel from './views/UserPage/SidePanels/MoveToTrashSidePanel';
-import FilterUsersSidePanel from './views/UserPage/SidePanels/FilterUsersSidePanel';
-import AssignCoachesSidePanel from './views/UserPage/SidePanels/AssignCoachesSidePanel';
-import RemoveFromClassSidePanel from './views/UserPage/SidePanels/RemoveFromClassSidePanel';
-import EnrollLearnersSidePanel from './views/UserPage/SidePanels/EnrollLearnersSidePanel';
+import MoveToTrashSidePanel from './views/users/sidePanels/MoveToTrashSidePanel';
+import FilterUsersSidePanel from './views/users/sidePanels/FilterUsersSidePanel';
+import AssignCoachesSidePanel from './views/users/sidePanels/AssignCoachesSidePanel';
+import RemoveFromClassSidePanel from './views/users/sidePanels/RemoveFromClassSidePanel';
+import EnrollLearnersSidePanel from './views/users/sidePanels/EnrollLearnersSidePanel';
 
 const logging = logger.getLogger(__filename);
