+++ kolibri/plugins/facility/assets/src/views/users/sidePanels/MoveToTrashSidePanel.vue
