+++ kolibri/plugins/facility/assets/src/views/users/sidePanels/RemoveFromClassSidePanel.vue
